update inventory, supplier

+++ backend/stockIO_service/src/main/java/vn/hcmuaf/edu/vn/stockio_service/dto/StockInItemDTO.java
@@ -3,16 +3,16 @@ package vn.hcmuaf.edu.vn.stockio_service.dto;
 import java.math.BigDecimal;
 
 public class StockInItemDTO {
-    private int productId;
+    private Long productId;
     private Integer quantity;
     private BigDecimal unitPrice;
     private String note;
 
-    public int getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(Long productId) {
         this.productId = productId;
     }
